requirements.txt
Flask
flask_cors==3.0.8
pytest==6.2.5
flask-sqlalchemy==2.5.1
psycopg2==2.9.1